# docs/swagger.yaml
basePath: /api
definitions:
  model.AddressResponse:
    properties:
      city:
        type: string
      country:
        type: string
      created_at:
        type: integer
      id:
        type: string
      postal_code:
        type: string
      province:
        type: string
      street:
        type: string
      updated_at:
        type: integer
    type: object
...
  model.CreateAddressRequest:
    properties:
      city:
        maxLength: 255
        type: string
      country:
        maxLength: 100
        type: string
      postal_code:
        maxLength: 10
        type: string
      province:
        maxLength: 255
        type: string
      street:
        maxLength: 255
        type: string
    type: object
...
  model.UpdateAddressRequest:
    properties:
      city:
        maxLength: 255
        type: string
      country:
        maxLength: 100
        type: string
      postal_code:
        maxLength: 10
        type: string
      province:
        maxLength: 255
        type: string
      street:
        maxLength: 255
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: RESTful API for managing users, contacts, and addresses with clean
    architecture principles.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Golang Clean Architecture API
  version: "1.0"
paths:
...
  /contacts/{contactId}/addresses:
    get:
      consumes:
      - application/json
      description: Get all addresses for a specific contact
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of addresses
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.AddressResponse'
                type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              errors:
                type: string
            type: object
        "404":
          description: Contact not found
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List addresses
      tags:
      - addresses
    post:
      consumes:
      - application/json
      description: Create a new address for a specific contact
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      - description: Address creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateAddressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created address
          schema:
            properties:
              data:
                $ref: '#/definitions/model.AddressResponse'
            type: object
        "400":
          description: Invalid request body
          schema:
            properties:
              errors:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              errors:
                type: string
            type: object
        "404":
          description: Contact not found
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new address
      tags:
      - addresses
  /contacts/{contactId}/addresses/{addressId}:
    delete:
      consumes:
      - application/json
      description: Delete a specific address by ID for a contact
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      - description: Address ID
        in: path
        name: addressId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted address
          schema:
            properties:
              data:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              errors:
                type: string
            type: object
        "404":
          description: Address not found
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an address
      tags:
      - addresses
    get:
      consumes:
      - application/json
      description: Get a specific address by ID for a contact
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      - description: Address ID
        in: path
        name: addressId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Address details
          schema:
            properties:
              data:
                $ref: '#/definitions/model.AddressResponse'
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              errors:
                type: string
            type: object
        "404":
          description: Address not found
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get an address
      tags:
      - addresses
    put:
      consumes:
      - application/json
      description: Update a specific address by ID for a contact
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      - description: Address ID
        in: path
        name: addressId
        required: true
        type: string
      - description: Address update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateAddressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated address
          schema:
            properties:
              data:
                $ref: '#/definitions/model.AddressResponse'
            type: object
        "400":
          description: Invalid request body
          schema:
            properties:
              errors:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              errors:
                type: string
            type: object
        "404":
          description: Address not found
          schema:
            properties:
              errors:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              errors:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an address
      tags:
      - addresses
...
securityDefinitions:
  BearerAuth:
    description: 'API token authentication. Format: your-token-here (without "Bearer"
      prefix)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"